First step to openstack provisioner
