+++ molecule/Provisioners/proxmoxprovisioner.py
@@ -1,0 +1,6 @@
+import baseprovsioner as BaseProvisioner
+
+# Place holder for Proxmox, partially implemented
+class ProxmoxProvisioner(BaseProvisioner):
+    def __init__(self):
+        super(ProxmoxProvisioner, self).__init__()
